[REFACTOR] loader, query 파일로 분리
