+++ client/src/components/community/CommunityMyPosts.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { useQuery } from '@tanstack/react-query';
 import { Chip, Flex, Group, List, Text } from '@mantine/core';
-import { communityMeQuery } from '../../pages/CommunityMe';
+import { communityMeQuery } from '../../query';
 import { EmptyPostIndicator, PostItem } from '.';
 
 const MyPosts = styled(List)`
