+++ client/src/loader/index.js
@@ -1,0 +1,5 @@
+export { default as communityCategoryLoader } from './communityCategoryLoader';
+export { default as communityPostLoader } from './communityPostLoader';
+export { default as communityMeLoader } from './communityMeLoader';
+export { default as rankLoader } from './rankLoader';
+export { default as profileLoader } from './profileLoader';
