+++ client/src/loader/rankLoader.js
@@ -1,0 +1,10 @@
+import { rankQuery } from '../query';
+
+const rankLoader = queryClient => async () => {
+  const query = rankQuery();
+
+  // eslint-disable-next-line no-return-await
+  return queryClient.getQueryData(query.queryKey) ?? (await queryClient.fetchQuery(query));
+};
+
+export default rankLoader;
