+++ client/src/pages/CommunityMe.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from '@emotion/styled';
 import { Container, Title } from '@mantine/core';
-import { getMyPosts } from '../api/posts';
+
 import { CommunityMyPosts, CommunitySkeleton } from '../components/community';
 
 const Wrapper = styled(Container)`
@@ -15,23 +15,6 @@ const Wrapper = styled(Container)`
   color: var(--font-color);
 `;
 
-const staleTime = 3000;
-
-const communityMeQuery = () => ({
-  queryKey: ['communityMe'],
-  queryFn: async () => {
-    const { data } = await getMyPosts();
-    return data;
-  },
-  staleTime,
-});
-
-const communityMeLoader = queryClient => async () => {
-  const query = communityMeQuery();
-  // eslint-disable-next-line no-return-await
-  return queryClient.getQueryData(query.queryKey) ?? (await queryClient.fetchQuery(query));
-};
-
 const CommunityMe = () => (
   <Wrapper>
     <Title size="52px" mt="40px">
@@ -43,5 +26,4 @@ const CommunityMe = () => (
   </Wrapper>
 );
 
-export { communityMeQuery, communityMeLoader };
 export default CommunityMe;
