+++ client/src/pages/CommunityPost.jsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 import { useParams } from 'react-router-dom';
 import { useQuery } from '@tanstack/react-query';
 import { Container, Divider } from '@mantine/core';
-import { getPost } from '../api/post';
+import { communityPostQuery } from '../query';
 import Comments from '../components/community/Comments';
 import PostContent from '../components/community/PostContent';
 
@@ -17,27 +17,6 @@ const Wrapper = styled(Container)`
   font-size: 0.75rem;
   color: var(--font-color);
 `;
-
-const staleTime = 3000;
-
-const communityPostQuery = postId => ({
-  queryKey: ['communityPost', postId],
-  queryFn: async () => {
-    const { data } = await getPost(postId);
-    return data;
-  },
-  staleTime,
-});
-
-export const communityPostLoader =
-  queryClient =>
-  async ({ params }) => {
-    const query = communityPostQuery(params.postId);
-
-    // eslint-disable-next-line no-return-await
-    const { post } = queryClient.getQueryData(query.queryKey) ?? (await queryClient.fetchQuery(query));
-    return post;
-  };
 
 const CommunityPost = () => {
   const params = useParams();
