+++ client/src/pages/index.js
@@ -8,3 +8,4 @@ export { default as RegisterProduct } from './RegisterProduct';
 export { default as CommunityPost } from './CommunityPost';
 export { default as Question } from './Question';
 export { default as Rank } from './Rank';
+export { default as CommunityCategory } from './CommunityCategory';
