+++ client/src/query/index.js
@@ -1,0 +1,5 @@
+export { default as categoryQuery } from './categoryQuery';
+export { default as communityPostQuery } from './communityPostQuery';
+export { default as communityMeQuery } from './communityMeQuery';
+export { default as rankQuery } from './rankQuery';
+export { default as profileQuery } from './profileQuery';
